Updated list api.

+++ src/rigid_list_internal.h
@@ -5,9 +5,8 @@
 
 #include <stdbool.h>
 
-void _rg_List_ShiftElements(struct rg_List *list,
-                            unsigned index, unsigned length,
-                            int shift_offset);
+void _rg_List_ShiftElements(struct rg_List *list, unsigned index,
+                            size_t length, int shift_offset);
 bool _rg_List_Resize(struct rg_List *list);
 
 #endif /* RIGID_LIST_UTILS_H */
